RK FB: make screen-centered scaling optional
This code forced a resized LCDC layer (window) to be placed in
the center of the screen. This removes a lot of the flexility of
the hardware which allows almost arbitrary placement and scaling
of a layer. It might have been used for the Android feature which
allows scaling down of the active output relative to the screen.
The now optional code must be disabled for hardware accelerated
XV support.