RK FB: add ioctl for reading fb addresses and length
Adds an ioctl which returns yrgb and cbr framebuffer addresses
and the length of the framebuffer. As far as I'm aware, this
information was only exposed through the sysfs interface, which
isn't very easy to use programmatically.

This information is used to allocate off-screen framebuffer memory
for temporary RGA buffers and for the video layer.