RK FB: Add option for 32BPP color depth
Sets the framebuffer to 32 bit BGRA color mode. Only enable on a
machine which has at least a 32 bit wide memory interface; this is
known to cause significant performance degradation on machines
with a 16 bit memory interface. This does not have any effect on
video displayed using the hardware accelerated XV driver.